edit wrap, avatar paddings, fix safe area listView
