+++ lib/constants/colors.dart
@@ -3,6 +3,7 @@ import 'dart:ui';
 class AppColors {
   const AppColors._();
 
+  static const white = Color(0xFFFFFFFF);
   static const iconGreen = Color(0xFF08a652);
   static const dividerGreen = Color(0xFF068441);
   static const black = Color(0xFF000000);
