+++ lib/ui/profile/profile.dart
@@ -1,17 +1,17 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_screen/ui/profile/carousel/carousel.dart';
-import 'package:flutter_screen/ui/profile/carousel/subTextSection.dart';
-import 'package:flutter_screen/ui/profile/filters/chips.dart';
-import 'package:flutter_screen/ui/profile/filters/textSection.dart';
-import 'package:flutter_screen/ui/profile/list/listTiles.dart';
-import 'package:flutter_screen/ui/profile/list/tarifsText.dart';
+import 'package:flutter_screen/ui/profile/carousel.dart';
+import 'package:flutter_screen/ui/profile/chips.dart';
+import 'package:flutter_screen/ui/profile/listTiles.dart';
+import 'package:flutter_screen/ui/profile/subTextSection.dart';
+import 'package:flutter_screen/ui/profile/tarifsText.dart';
+import 'package:flutter_screen/ui/profile/textSection.dart';
 
 class profileScreen extends StatelessWidget {
   const profileScreen({super.key});
 
   @override
   Widget build(BuildContext context) {
-    return ListView(children: const [
+    return ListView(padding: EdgeInsets.zero, children: const [
       subscriptions(),
       carousel(),
       tarifksText(),
