+++ lib/ui/profile/tarifsText.dart
@@ -5,8 +5,8 @@ import 'package:flutter_screen/constants/sizes.dart';
 import 'package:flutter_screen/constants/strings.dart';
 
 // ignore: camel_case_types
-class subscriptions extends StatelessWidget {
-  const subscriptions({super.key});
+class tarifksText extends StatelessWidget {
+  const tarifksText({super.key});
 
   @override
   Widget build(BuildContext context) {
@@ -16,20 +16,20 @@ class subscriptions extends StatelessWidget {
       children: [
         Padding(
             padding: EdgeInsets.only(
-                bottom: 8,
-                left: AppPaddings.mainSidePadding,
-                right: AppPaddings.mainSidePadding),
-            child: Text(Strings.subTitleSection,
+                top: AppPaddings.tarifsTextTop,
+                bottom: AppPaddings.tarifsTextBottom,
+                left: AppPaddings.mainSidePadding),
+            child: Text(Strings.tarifsAndLimits,
+                textAlign: TextAlign.left,
                 style: TextStyle(
                     fontSize: Sizes.fontTitleSize,
                     fontWeight: FontWeight.w700,
                     color: AppColors.black,
                     height: Sizes.heightTitle))),
         Padding(
-            padding: EdgeInsets.only(
-                left: AppPaddings.mainSidePadding,
-                right: AppPaddings.mainSidePadding),
-            child: Text(Strings.subTextSection,
+            padding:
+                EdgeInsets.only(left: AppPaddings.mainSidePadding, bottom: 12),
+            child: Text(Strings.secondsubTarif,
                 textAlign: TextAlign.left,
                 style: TextStyle(
                   fontSize: Sizes.fontRegularSize,
